requirements.txt
librosa>=0.10.1
noisereduce>=2.0.1
soundfile>=0.12.1
numpy>=1.22.0
pymongo>=4.6.0
motor>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0 